pyproject.toml
[project]
name = "dnd-story-illustrator"
version = "0.1.0"
description = "DnD Story Illustrator: characters, scenes, and session montages with OpenAI"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "streamlit>=1.36.0",
  "openai>=1.40.0",
  "Pillow>=10.3.0",
  "reportlab>=4.2.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
]

[project.scripts]
dnd-illustrator = "app.main:main"

[tool.uv]
dev-dependencies = []


